set max TLS version to 1.2 for internal services
We want to prevent library consumers from being surprised by the TLS 1.3
bump in Go 1.13. There shouldn't be any issues with TLS 1.3 as defined
in the spec but we have been receiving feedback on issues which we think
are related to TLS 1.3:

* Requests between consumers of this library and Java are showing higher
  load.

* Higher CPU usage when stressing a Go server.

We should be able to increase this maximum version to 1.3 after some
further vetting of major consumers.

Co-authored-by: Christopher Brown <cbrown@pivotal.io>