Remove vendor directory and add .gitignore
Removes vendored dependencies because this module doesn't have a main
package, and is typically imported as a library. Therefore, the
importing code should vendor the code if they so desire.

Adds a .gitignore file (straight from the github .gitignore Go template)
that prevents the vendor directory from being added back in the future.